Add a GPS indicator to the camera
This will add a GPS indicator to the preview of the photo camera. It is placed as an overlay at the upper top corner of the photo camera preview.
It shows one of three icons depending on the GPS state: acquiring fix, GPS fix acquired and GPS disallowed in system settings.

The icon will only be shown in the preview of the photo module, not in the video or panorama module, and not in the gallery view.

Patch Set 1: use cm_strings.xml and cm_arrays.xml instead of strings.xml and arrays.xml

Drawables by blunden.